requirements.txt
streamlit
pypdf
langchain
python-dotenv
torch
accelerate
bitsandbytes
transformers
sentence_transformers
faiss_cpu
chainlit
ctransformers
typer